Remove dependency cycle
